TOB-6019 refactor bestillingbatcher

+++ src/main/kotlin/no/nav/sokos/skattekort/module/skattekort/BestillingRepository.kt
@@ -23,7 +23,7 @@ object BestillingRepository {
 
     fun getBestillingsKandidaterForBatch(
         tx: TransactionalSession,
-        maxYear: Int,
+        maxYear: Short,
     ): List<Bestilling> =
         tx.list(
             queryOf(
